requirements.txt
torch
torchvision
Pillow
numpy
opencv-python
scipy
onnxruntime
matplotlib
spandrel
color-matcher
